alertlib ========
Summary:
A small library to make it easy to send alerts to various platforms.

This library consists of a python module that will let you send an
alert to any or all of

   * HipChat
   * PagerDuty
   * khanacademy.org email lists
   * GAE logs and/or syslog

You must provide a decrypted secrets.py (from the webapp repo) to use
these services.

Test Plan:
I manually tested the HipChat.  I'll test the others once this is
checked in, on an ec2 machine that can send email.  I'll maybe also
try making a small appengine app to test that too.

Reviewers: chris, ben

Differential Revision: http://phabricator.khanacademy.org/D8561