Be careful to check that syslog exists when assigning _LOG_TO_SYSLOG.
When I hoisted _LOG_TO_SYSLOG out of the send_to_logs, I hoisted it
out of the try/except that dealt with syslog not being importable (as
indeed it isn't from within the appengine sandbox).

Now I use a separate try/except for that.

Auditors: chris

Test Plan:
python
>>> import sys
>>> sys.modules['syslog'] = None
>>> import alertlib

That fails before this change but succeeds after.

I also added a testcase to alertlib_test.py, though it doesn't check
things as well because the imports succeed the first time around.