Work correctly without secrets.
We need this so we can use alertlib in webapp tests, without requiring
the people running tests to have secrets involved.  In any case, we
don't want tests to actually write to hipchat/etc!

This is not a perfect solution: we could still send mail via smtp even
in test mode, and if you run tests and *do* have secrets, you may find
yourself writing to hipchat.  But it's better than before.

Auditors: chris

Test Plan:
Deleted secrets.py and ran python and did this:

    >>> import alertlib
    >>> alertlib.Alert('test message') \
    ...                 .send_to_hipchat('1s and 0s') \
    ...                 .send_to_email('ka-admin') \
    ...                 .send_to_pagerduty('oncall') \
    ...                 .send_to_logs() \
    ...                 .send_to_graphite('stats.alerted')
    WARNING:root:Not sending this to hipchat (no token found): {'from': 'AlertiGator', 'message_format': 'text', 'color': 'purple', 'room_id': '1s and 0s', 'notify': 0, 'message': 'test message'}
    WARNING:root:Not sending to graphite; no API key found: stats.alerted 1

I *did* see it trying to send email via smtp (due to the fact that I
didn't have google mail working properly), which is a TODO to do
better with.