Properly handle the case when syslog is not available
Summary:
In production, on App Engine where syslog isn't available, we were seeing
KeyErrors where _mapped_severity() was trying to do severity_map[logging.INFO]
as a default but severity_map was an empty dictionary. I'm pretty sure the
except/pass check was supposed to catch that, but it was looking for a
NameError instead of a KeyError.

Test Plan:
Fingers crossed, I'm not sure about the best way to disable syslog
locally to trigger this.

Reviewers: chris, csilvers

Reviewed By: csilvers

Subscribers: csilvers

Differential Revision: http://phabricator.khanacademy.org/D8825