Allow a custom sender in HipChat notifications
Summary:
While "AlertiGator" is plenty clever, our general alerting framework shouldn't
be restricted to just that when there are so many other members of the animal
kingdom that could potentially take part in the alerting fun.

Test Plan:
I reran the manual test from D8950 but with "BigBingo Baboon" as a custom
sender. It worked.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: ben

Differential Revision: http://phabricator.khanacademy.org/D8951