Don't let failures to send to hipchat raise a 500 error.
If we fail to send for any reason, we just log and continue.  Before
we were only doing that if we got back a non-200; a connection-timeout
or something would still raise an exception.

I also fixed a bug where we had stopped catching NameError for
syslog.  We need to catch that *in addition to* KeyError.

Auditors: chris

Test Plan:
tests/alertlib_test.py