add slack support to AlertLib
Summary:
Adds Slack support to AlertLib and adds a much more detailed HOWTO to
the README in order to assist people in transitioning to Slack.

A generic `send_to_chat` wrapper was intentionally omitted, as a.)
chaining fits the existing model better, b.) it is better to be aware
of the differences between the two destinations instead of trying to
paper them over, and c.) sending to both simultaneously is an edge case
only temporarily used in transitions, where having explicit control is
actually preferable. (This is a reversal of my original opinion after
spending a few days with the code).

This will need a corresponding update to secrets.py in webapp (before
being landed?).

Test Plan:
- `make tests`
- use alert.py command line client

Reviewers: emily, benjaminpollack, ethan

Reviewed By: benjaminpollack, ethan

Differential Revision: https://phabricator.khanacademy.org/D20446