Use Slack's default sender info rather than our own
Summary: Rather than setting our own default `username` and `icon_emoji`/`icon_url`, just don't pass them unless set, which will use whatever default is set up in the Slack incoming webhook for alertlib.  With a config change in Slack (setting the sender to be `AlertiGator` and the icon to be `:crocodile:`), this will keep the behavior the same, except that `icon_url` will now work as expected, rather than being overridden on the Slack side by our default `icon_emoji`.  I'll make the config change before deploying this.

Test Plan: Sent alerts from an ipython shell and from `alert.py` with and without a sender, and with and without an `icon_emoji` or `icon_url`.

Reviewers: benjaminpollack, mroth, csilvers

Reviewed By: mroth, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21806