send_to_asana followers parameter now takes user emails instead of ids
Summary:
Updates send_to_asana followers paramaeter to take user emails and map the
emails to user ids in a similar method to tags and projects.

Test Plan:
Two new unit tests were added to alertlib test to test looking up followers
in the user cache. A semi-manual, uncommitted test was also used to test integration with Asana, and it was observed that a task with followers was created.

Reviewers: benkraft, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D28003