Remove trailing colon in Asana task name when necessary
Summary: Asana task names ending with a colon became section headings. The last character of the name is now removed if it is a colon.

Test Plan:
Added a unit test to test that a trailing colon is removed from the Asana task name. Also tested sending an alert to asana with a trailing colon and observed
that the name added to asana was not a section heading.

Reviewers: csilvers, benkraft

Reviewed By: csilvers, benkraft

Differential Revision: https://phabricator.khanacademy.org/D28359