Refactor the send-datapoint method to make it easier to mock.
I want to mock it out in some client code to allow me to send a bunch
of datapoints in one alertlib call.  This is what's necessary to do
that (mostly just passing in the list instead of listifying in the
function).

Auditors: dhruv

Test Plan:
tests/alertlib_test.py