Add the ability to override the timestamp for stackdriver, and re-refactor.
I meant to include the re-refactoring in my previous audit, but forgot
to do a `git amend`, so I'm doing it here. :-(

The addition of `when` will be useful when using this to send
historical data.  Maybe I am making the stackdriver call too
complicated.

Auditors: dhruv

Test Plan:
tests/alertlib_test.py