No longer allow passing in the kind (GAUGE vs CUMULATIVE).
Summary:
It turns out that the stackdriver API ignores the setting of the kind:
https://cloud.google.com/monitoring/api/ref_v3/rest/v3/TimeSeries says
about `metricKind`:
> This field is ignored when writing data; the value specified in the
> descriptor is used instead. @OutputOnly

Instead, timeSeries figures out the metricKind by examining the
metricDescriptor for the metric you are passing into the time series.

So we no longer bother to pass in `metricKind`.  If you want to create
a CUMULATIVE metric, you'll need to create a metricDescriptor for that
metric manually first, which has the right type.

Test Plan:
tests/alertlib_test.py
tests/timeout_test.py

Reviewers: chris

Reviewed By: chris

Subscribers: alexanderforsyth

Differential Revision: https://phabricator.khanacademy.org/D28743