Try again if we get an INVALID_ARGUMENT.
Sometimes the cloud-monitoring API says this:
```
{
  "error": {
    "code": 400,
    "message": "One or more TimeSeries could not be written.",
    "errors": [
      {
        "message": "One or more TimeSeries could not be written.",
        "domain": "global",
        "reason": "badRequest"
      }
    ],
    "status": "INVALID_ARGUMENT"
  }
}
```

Our working theory is that this is because the timestamp that we give
is, from google's point of view, in the future.  This can happen
because we give `time.time()` as the timestamp, and if google is a bit
behind us they will get flummoxed.  Just retrying after a short wait
should fix things.

This check may catch other problems that refixing *won't* help, since
it's a bit broad, but that's ok; at worst it's just a bit of
unnecessary extra work.

Auditors: chris

Test plan:
Fingers crossed