Split alertlib into multiple files.
Summary:
The `__init__.py` file was getting really large, and it separates very
nicely since each `send_to_foo` function is its own thing, with very
little shared code.  So now I split it up, one file per send_to_foo.

The only hard part was the tests, which did a lot of mocking that I
had to redo.

Move as many methods as we can to free functions.

One problem with mixins is name collisions: if MixinA defines a method
called `_post_data` and so does MixinB, then only one will win, with
bad results.

The easiest fix is to just not have methods in your mixin.  So I do
that where it's practical, which is most of the time.  Now most mixins
just define `send_to_foo`.  Where it's not practical to use a free
function, I make sure the helper functions have the service name in
them.

Test Plan:
   env PYTHONPATH=${HOME}/khan/webapp/third_party/frankenserver/lib/apiclient:${HOME}/khan/webapp/third_party/frankenserver/lib/oauth2client:${HOME}/khan/webapp/third_party/frankenserver/lib/httplib2:${HOME}/khan/webapp/third_party/frankenserver/lib/uritemplate tests/alertlib_test.py
   env PYTHONPATH=${HOME}/khan/webapp/third_party/frankenserver/lib/apiclient:${HOME}/khan/webapp/third_party/frankenserver/lib/oauth2client:${HOME}/khan/webapp/third_party/frankenserver/lib/httplib2:${HOME}/khan/webapp/third_party/frankenserver/lib/uritemplate tests/timeout_test.py

(same)

Reviewers: chris

Reviewed By: chris

Subscribers: #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D32988