Make alertlib python 2/3 compatible
Summary:
Alertlib is now a dependency for the Khanalytics pipelines library, which
is written in Python3, so we need Alertlib to be python3 compatible.  Of
course, webapp is still on python 2.7, so everything now works for both
versions of python

Also did some refactoring to handle string encoding more consistently.

Only changed the behavior on email string encoding, and only for Python 3.
There was a test to make sure that in python2, a message with utf-8 characters
still sent an ascii-encoded email, and I'm not sure why we want that, but I
made sure the behavior is still the same.  However in Python 3, if there's
utf-8 in the message, we'll go ahead and send the email with utf-8.

Test Plan:
all tests pass, whether they're run in python 2.7 or Python 3.6.
Also confirmed that emails with utf-8 send and render properly.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33075