Ignore stackdriver timeseries errors due to old data.
Summary:
Ignore stackdriver errors that occur when re-uploading data.

In practice this occurs when retrying after there is an internal
error in stackdriver when uploading data. Some of the data is written
during the failed request, so we get an error when retrying with the
same data.

Also document requirements. The libary requires six and the tests
require a couple libraries listed in dev_requirements.txt

https://app.asana.com/0/198245187747885/249999927560406

Test Plan:
- bin/python -m unittest discover -p '*_test.py' tests

- Create an empty virtualenv and just install the things
  in dev_requirements.txt and make sure the tests run.

Reviewers: csilvers, dhruv

Reviewed By: dhruv

Differential Revision: https://phabricator.khanacademy.org/D33149