Integrate new Alerta mixin to receive send_to_alerta requests and call alerta API
Summary:
This mixin sends alerts to Alerta--a monitoring system used to consolidate
and de-duplicate alerts from multiple sources. It should be called from any
cron jobs or various monitoring tools with the goal of providing the Alerta
dashboard with a snapshop of what's broken at any given time. With KA
permissions, this dashboard can be viewed on https://alerta.khanacademy.org.

Within the cron job or other origin, this mixin must be called using not
only --aggregator, but also the additional args --aggregator_resource and
--aggregator_event_name. All are required in order for an alert to be
sent through successfully.

Note: The flag --aggregator was used over simply --alerta so as to make these
calls more generalizable in the case that we ever need to shift to a different
aggregator in the future. This way the code where these calls originate does
not need to be updated, only adaptation of alertlib would be required.

Changes include:
- Build alerta mixin for receiving alerts objects and sending to Alerta dashboard
- Integrate this new mixin with Alert class and alert.py script
- Update mixin to handle missing args
- Fix bugs with Alerta API call and api_key handling
- Add tests for new mixin

Test Plan:
- Run alertlib_test.py
- Successfully sends test alerts to Alerta by running script in the CLI.

Reviewers: nabil, drosile, csilvers

Reviewed By: nabil

Differential Revision: https://phabricator.khanacademy.org/D36645