Remove ka_secrets hack
With the addition of specifying a directory to look in for secrets, when we
substate this change into webapp we can just set that variable, and get rid of
this hack today.  The future is now!

Test plan: Substate into webapp; alertlib itself never used this hack anyway.

Auditors: csilvers