Allow secrets to not exist even if a directory is specified
Unfortunately, Jenkins seems to be leaking environment variables in places we
don't want it to be, so ALERTLIB_SECRETS_DIR is set but there are no secrets in
it.  (I haven't bothered to work out the details of exactly when/why this
happens.)  Ideally we might fix Jenkins, but it's simple enough to fall back to
no secrets even if the environment variable is set, so we'll just do that.

Test plan: substate into webapp, and deploy it.

Auditors: csilvers