Sanitize auth information when logging stackdriver errors.
Summary:
I did an audit of all the places we log in alertlib, and this is the
only one that logs something that might plausibly have sensitive
headers in it.  (In other places we log requests, but they don't log
header info it doesn't look like.)

The other place we have to worry is slack, where (it seems) the auth
token is part of the url, which gets logged.  But I'm less worried
about our slack integration token leaking -- if indeed that's even a
risk here -- since those are so easily revoked.

Fixes https://app.asana.com/0/25429045350268/185283653974479

Test Plan: python tests/alertlib_test.py

Reviewers: colin

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D39208