Add ability to send informational alerts to alerta
Summary:
As part of the the Log of What's Happened Recently
project, we want the ability to send informational
alerts to alerta that will expire after a short
period of time.

https://app.asana.com/0/412860343989559/421978121271778
https://app.asana.com/0/412860343989559/426450392536571

This diff adds two arguments to the `send_to_alerta`
method:

- resolve, which is used to explicitly indicate whether
  an alert should be resolved. For now it's set to resolve
  info alerts by default, but will be changed later to not
  resolve by default once callers are updated to explicitly
  specify resolve.

- timeout, which overrides the default alerta timeout.

Used together you can create an informational (green)
alerta that expires after a short period of time:

    a = alertlib.Alert('Some information')
    a.send_to_alerta('infrastructure', resource='test', event='event', resolve=False, timeout=60*5)

Note: for the timeout to work, a cron job must be running
which expires alerts.
https://docs.google.com/document/d/16yL6OoqAT2GpCC7uA-59CfPkD0k7VFAAReXhWixl4GA/edit

For now resolve defaults to True. Once all the callers
of this method are updated to pass True, the default can
be changed to False. When that happens alerts with a severity
of logging.INFO will not auto-close when sent to alerta unless
you specify `resolve=True`.

Test Plan:
- Manually create an alert with severity of logging.INFO.
  Send it to alerta and confirm that it is auto-closed since
  it has a severity of "cleared".

- Send the alert to alerta again (with a different event)
  but specify `resolve=False`
  and confirm that it isn't auto closed and has a severity
  of informational.

- Send an to alerta with a short timeout and confirm that
  the alert is expired after the timeout number of seconds.

- Manually test the cli:
```
    bin/python alert.py --aggregator-resource=test --aggregator-event=e1 --aggregator=infrastructure
    bin/python alert.py --aggregator-resource=test --aggregator-event=e1 --aggregator=infrastructure --aggregator-no-resolve
    bin/python alert.py --aggregator-resource=test --aggregator-event=e2 --aggregator=infrastructure --aggregator-no-resolve --aggregator-timeout=20
```

Reviewers: nabil, jacqueline, csilvers

Reviewed By: nabil, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D39296