Fix a double encoding issue in the alertlib notifier
Summary:

In 72b6fd37d1cc1c5bff228a84ea617064256f1912, I fixed an issue where we
were sending an inconsistent mix of bytes and strings to `json.dumps`
by ensuring that everything we were passing to that function was a
string and then encoding it to bytes.

However, when I tried this talking to the production alerta service, I
got another error because in the post request to alerta, we're encoding
the json again! This means that we were double-encoding the message
text prior to any of my encoding fixes, and after
72b6fd37d1cc1c5bff228a84ea617064256f1912, we're now double-encoding the
whole payload.

This commit removes the encode I added in
72b6fd37d1cc1c5bff228a84ea617064256f1912 so that we're just encoding
the payload once.

Test Plan:
- in python2 `python -m unittest discover -p '*_test.py' tests`
- hack khanalytics a bit so that I can use my local copy of alertlib,
  but now with prod alerta credentials installed, verify that now when
  a pipeline fails in khanalytics, messages make it all the way to
  alertlib

Auditors: jacqueline, amos