Fix tests on python 3
Summary:
We had a bunch of tests failing on python 3.  (Some may have been new
with 3.6?  I'm not sure.)  Now we don't.  I made a
few changes:
- I disabled the email tests python 3, because the parts requiring
  google mail don't work (due to the lack of a python 3 appengine
  environment) and those weren't easily separable (although I left a TODO
  to separate them).
- In several tests, I added `b'`ytestrings where appropriate, so we get
  the types right in python 3.
- In one case, I changed an assertion slightly so as not to assert about
  dict iteration order, which changed in 3.6.
The timeout tests are also logging some warnings in Python 3.6, which I
decided not to worry about, but left a TODO.

Test Plan:
ran tests on both python 3.6.3 and python 2.7.14.  There is not
currently a single command to do this, which I intend to fix in a
separate commit.

Reviewers: csilvers, amos, nabil, colin

Reviewed By: colin

Differential Revision: https://phabricator.khanacademy.org/D42187