Add support for modern oauth2clients in addition to the old one we use.
Summary:
We want to update oauth2client in webapp, which means alertlib has to
support the modern API.  It turns out it's a very simple change to do
that (I had expected it to be much harder!)

I continue to support the old API as well, for clients (e.g. ec2
machines) that don't want to update their oauth2client just yet.

Test Plan:
I ran, in webapp-root:
   % env ALERTLIB_SECRETS_DIR=shared python
   >>> import third_party.alertlib.stackdriver as s
   >>> s.send_datapoints_to_stackdriver({})
This gave back a 400 error, because I was not passing in a valid
metric.  But the authentication succeeded!  To validate that,
I changed a byte in the `google_alertlib_service_account` secret
in shared/secrets.py -- I changed an `e` to an `f` in `private_key` --
and did the test plan again, and this time I got an
`invalid_grant` Exception.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D42847