Allow sending in reply to a slack thread
Summary:
Slack has threads, which can be used to keep replies to a message
together.  It would be nice for alertlib to be able to send to them, to
keep the deploy channel a bit cleaner.  This commit adds support to
alertlib; you just pass `--slack-thread`.

Ideally there could be a similar feature for several targets: comments
on a bugtracker, replies to an email, and replies in slack could have
similar semantics, and alertlib could do all of them based on the same
settings.  But I didn't figure out a way to specify the message to which
we should reply in a platform-agnostic way.  So at present this is all
slack-specific.

Unfortunately, life was not quite so simple as just passing an extra
parameter with the message.  In addition, I had to add functionality to
use a slack bot API token instead of an incoming webhook; the two have
similar functionality for the most part but only the former supports
threads.  This results in some subtle semantics differences which I
tried to bridge to the extent possible.  I'm sure we'll discover more as
we go.  But I think bot tokens are generally better-supported so
hopefully once we move to them we'll be a bit happier in general.

Test Plan:
- send a message to `#bot-testing`, using
  https://api.slack.com/methods/chat.postMessage/test
- note the "ts" value, in this case "1523575279.000171"
- add a alertlib_slack_bot_token to secrets
- `alertlib.Alert("test reply").send_to_slack(channel='#bot-testing', thread='1523575279.000171')`
  should post a reply to that thread.
- `echo test reply | ./alert.py --slack '#bot-testing' --slack-thread '1523575279.000171'`
  similarly.
- test a few options (including `--chat-sender`, `--icon-emoji`, `--icon-url`) to make sure they work

Reviewers: amos, colin, csilvers, jacqueline

Reviewed By: amos, csilvers

Differential Revision: https://phabricator.khanacademy.org/D43796